@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class ConfirmationTest < ActionDispatch::IntegrationTest
+class ConfirmationTest < Devise::IntegrationTest
 
   def visit_user_confirmation_with_token(confirmation_token)
     visit user_confirmation_path(confirmation_token: confirmation_token)
@@ -184,14 +184,14 @@ class ConfirmationTest < ActionDispatch::IntegrationTest
 
   test 'resent confirmation token with valid E-Mail in XML format should return valid response' do
     user = create_user(confirm: false)
-    post user_confirmation_path(format: 'xml'), user: { email: user.email }
+    post user_confirmation_path(format: 'xml'), params: { user: { email: user.email } }
     assert_response :success
     assert_equal response.body, {}.to_xml
   end
 
   test 'resent confirmation token with invalid E-Mail in XML format should return invalid response' do
     create_user(confirm: false)
-    post user_confirmation_path(format: 'xml'), user: { email: 'invalid.test@test.com' }
+    post user_confirmation_path(format: 'xml'), params: { user: { email: 'invalid.test@test.com' } }
     assert_response :unprocessable_entity
     assert response.body.include? %(<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<errors>)
   end
@@ -213,7 +213,7 @@ class ConfirmationTest < ActionDispatch::IntegrationTest
   test 'request an account confirmation account with JSON, should return an empty JSON' do
     user = create_user(confirm: false)
 
-    post user_confirmation_path, user: { email: user.email }, format: :json
+    post user_confirmation_path, params: { user: { email: user.email }, format: :json }
     assert_response :success
     assert_equal response.body, {}.to_json
   end
@@ -249,7 +249,7 @@ class ConfirmationTest < ActionDispatch::IntegrationTest
   end
 end
 
-class ConfirmationOnChangeTest < ActionDispatch::IntegrationTest
+class ConfirmationOnChangeTest < Devise::IntegrationTest
   def create_second_admin(options={})
     @admin = nil
     create_admin(options)
