@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class PasswordTest < ActionDispatch::IntegrationTest
+class PasswordTest < Devise::IntegrationTest
 
   def visit_new_password_path
     visit new_user_session_path
@@ -160,8 +160,7 @@ class PasswordTest < ActionDispatch::IntegrationTest
     assert_response :success
     assert_current_url '/users/password'
     assert_have_selector '#error_explanation'
-    assert_contain Devise.rails4? ?
-      "Password confirmation doesn't match Password" : "Password doesn't match confirmation"
+    assert_contain "Password confirmation doesn't match Password"
     assert_not user.reload.valid_password?('987654321')
   end
 
@@ -251,14 +250,14 @@ class PasswordTest < ActionDispatch::IntegrationTest
 
   test 'reset password request with valid E-Mail in XML format should return valid response' do
     create_user
-    post user_password_path(format: 'xml'), user: {email: "user@test.com"}
+    post user_password_path(format: 'xml'), params: { user: {email: "user@test.com"} }
     assert_response :success
     assert_equal response.body, { }.to_xml
   end
 
   test 'reset password request with invalid E-Mail in XML format should return valid response' do
     create_user
-    post user_password_path(format: 'xml'), user: {email: "invalid.test@test.com"}
+    post user_password_path(format: 'xml'), params: { user: {email: "invalid.test@test.com"} }
     assert_response :unprocessable_entity
     assert response.body.include? %(<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<errors>)
   end
@@ -266,7 +265,7 @@ class PasswordTest < ActionDispatch::IntegrationTest
   test 'reset password request with invalid E-Mail in XML format should return empty and valid response' do
     swap Devise, paranoid: true do
       create_user
-      post user_password_path(format: 'xml'), user: {email: "invalid@test.com"}
+      post user_password_path(format: 'xml'), params: { user: {email: "invalid@test.com"} }
       assert_response :success
       assert_equal response.body, { }.to_xml
     end
@@ -275,8 +274,9 @@ class PasswordTest < ActionDispatch::IntegrationTest
   test 'change password with valid parameters in XML format should return valid response' do
     create_user
     request_forgot_password
-    put user_password_path(format: 'xml'), user: {
+    put user_password_path(format: 'xml'), params: { user: {
       reset_password_token: 'abcdef', password: '987654321', password_confirmation: '987654321'
+      }
     }
     assert_response :success
     assert warden.authenticated?(:user)
@@ -285,7 +285,7 @@ class PasswordTest < ActionDispatch::IntegrationTest
   test 'change password with invalid token in XML format should return invalid response' do
     create_user
     request_forgot_password
-    put user_password_path(format: 'xml'), user: {reset_password_token: 'invalid.token', password: '987654321', password_confirmation: '987654321'}
+    put user_password_path(format: 'xml'), params: { user: {reset_password_token: 'invalid.token', password: '987654321', password_confirmation: '987654321'} }
     assert_response :unprocessable_entity
     assert response.body.include? %(<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<errors>)
   end
@@ -293,7 +293,7 @@ class PasswordTest < ActionDispatch::IntegrationTest
   test 'change password with invalid new password in XML format should return invalid response' do
     user = create_user
     request_forgot_password
-    put user_password_path(format: 'xml'), user: {reset_password_token: user.reload.reset_password_token, password: '', password_confirmation: '987654321'}
+    put user_password_path(format: 'xml'), params: { user: {reset_password_token: user.reload.reset_password_token, password: '', password_confirmation: '987654321'} }
     assert_response :unprocessable_entity
     assert response.body.include? %(<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<errors>)
   end
@@ -301,7 +301,7 @@ class PasswordTest < ActionDispatch::IntegrationTest
   test "when using json requests to ask a confirmable request, should not return the object" do
     user = create_user(confirm: false)
 
-    post user_password_path(format: :json), user: { email: user.email }
+    post user_password_path(format: :json), params: { user: { email: user.email } }
 
     assert_response :success
     assert_equal response.body, "{}"
